Split out byte formatting into scriptUtil.ts
